pyproject.toml
[project]
name = "iie-qq-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.30.0",
    "matplotlib>=3.10.0",
    "nonebot-adapter-console>=0.10.0",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot2[fastapi,httpx]>=2.5.0",
    "sqlalchemy[asyncio]>=2.0.40",
]
